requirements.txt
Flask-Discord-Interactions==1.7.2
gunicorn==20.1.0
mysql-connector-python==8.0.29
python-i18n==0.3.9
PyYAML==6.0
